+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.6.1.alpha
+* the ajaxp:done event now sends an extra argument - the ajax_loadzone DOM element. This makes it easier for the event handler to make animations involving the loadzone. If it does not exist, null is returned.
+* one set of internally used data-pagination attributes has been renamed to data-ajax_section_id.
+* the ?pagination= querystring parameter has been renamed to ajax_section.
+
 ## v0.6.0
 The :pagination option no longer makes as much sense, now that the sections are called ajax_section. Also, this gem is as much about site navigation as pagination. This is the reason for some of the following changes.